New functionality: Adding a new user from the webpage
Added functionality to the earlier created form for adding a new user.
New users can only be created by users with admin privileges.
A request can be send by pressing the "OK" button or the Enter key.
A valid request will add the new and the hashed password to the database.

The Server will always return a status. Possible results:
 - 200 'OK'		-> The new user has been added to the system
 - 403 'Forbidden'	-> The user is not authorized to create a new user
 - 409 'Conflict'	-> The user already exists