FIX: indepentent paths, ADDED: loading animation example
FIXED:
Before, starting the node server from a different directory than /src caused errors while connecting to the database.
This was fixed by setting a const path to the basedirectory of the server.js /data/data.db
Also, the /public and /views folders are now independent paths as well by using app.use('/public'...) and app.set('views'...)

ADDED:
Added a simple example for a possible loading animation to the control panel. It can be triggered by pressing a button and uses an animated svg.