Target values: storing on server, initializing display
The target values for pressure and fan speed are now being stored on the server.
After loading the control panel page, the client will request the current target values and set the range inputs and labels to the correct values.
If the user changes the target value, the client will tell the server to update the values by sending a post request.