Added data minimization & soring of mode and target values
Data minimization:
The arrays of data fetched by the control panel to display the plots can contain a huge ammount of data. Sending and displaying a lot of data does not only increase the loading time of the control panel but also has a critical impact on the performance of the page itself. To prevent this from happening, the server now checks the ammount of data loaded and reduces it to a dataset with less than 2000 entries by removing every second entry (multiple times).
2000 datapoints do not have a critical impact on loading time or performance.